Change the world with tangible actions
INTRODUCTION

We are human beings, We work to deliver human development for all. humanity has been one of the most rewarding experiences of our life, we want to contribute to people's wellbeing and provide them with a better life, peace, human rights, and humanitarian efforts that are needed more than ever within Ukraine, Afghanistan, Yemen, Mexico, Syria, Haiti, Algeria, Iraq, DR Congo, South Sudan, Ethiopia and Sudan. Together we have power of our demands, peace and harmony are the foundation of human development.

Kyiv is the capital city of Ukraine. Ukraine is a country in Eastern Europe. Ukraine borders the Russian Federation to the east and North-East, Belarus to the North-West, Poland, Slovakia and Hungary to the West, Romania and Moldova to the South-West, and the Black Sea and Sea of Azov to the South and Southeast. Ukraine has a mostly temperate continental climate, although the southern Crimean coast has a humid subtropical climate. large nation that has a large amount of equipment and missiles starts a war and invades Ukraine due to a misunderstanding, according to President of Ukraine Volodymyr Zelensky said he doesn't consider diplomacy with Russia possible under the current conditions.

WHAT DO BELIEVE?

There is 8 billion people on the planet who is like planet to be health and wasteless, together and in solidarity we making possible for Kyiv to be safe, greener, cleaner in true harmony while many people think is now impossible. War is waste and is the main reason for the violation of human rights. We must stand stronger together and in solidarity to build a wasteless world in all form, together we have power of our demands, we need sustainable peace around the world, we need the health of our cities, we need the health of our nature, we need the health of our planet, we need the health of our people too! let's go together in harmony!

ONE ACTION AT A TIME WE CAN CHANGE THE KYIV?

The 2030 Agenda is an action plan for prosperity, planet and people too, implemented through the Sustainable Development Goals (SDGs). More and more of us are becoming aware of the need for sustainable peace and harmony. Of course we need resources and we need inclusions, inclusions from different voice, groups, indigenous people and community, together we was raise the Sustainable Development Goals flag but together we have power to build the map of Sustainable Development Goals (SDGs) in every corner of the world.

OUR MESSAGE?

Human development can not happen without dedicated environmental sustainability, building sustainable peace and equitable development. There is nothing more important than human dignity and no one has right to violate it. 

Diversity, inclusion, justice, equity and Open Data can help improve smart cities and urban infrastructure. It can also improve disaster resilience and ensure that critical resources are deployed effectively in an emergency. 

United People and INDUSTRY 5.0 remains on the front lines to build WasteLess world by systematic waste prevention and responding to conflicts and human rights violations. We believe war is waste so there will be no waste if we don't create waste, let's go together to build a new Kyiv in true harmony with the ecosystem of which we too are an important part